Fix wrong indexes returned in resolve[type] of ShakeInterpreterPackage in some cases
